+++ include/rectangle.h
@@ -7,15 +7,15 @@ namespace glvis {
 
     class Rectangle : public Shape {
     public:
-        Rectangle(double width, double height);
-        double getWidth() const;
-        double getHeight() const;
+        Rectangle(float width, float height);
+        float getWidth() const;
+        float getHeight() const;
         Vector2 getSize() const;
         void render() const override;
 
     private:
-        double width = 0.0;
-        double height = 0.0;
+        float width = 0.0f;
+        float height = 0.0f;
         unsigned int VAO = 0;
         unsigned int VBO = 0;
         unsigned int EBO = 0;
