+++ include/shape.h
@@ -6,22 +6,22 @@ namespace glvis {
 
     class Shape {
     public:
-        double getX() const;
-        double getY() const;
+        float getX() const;
+        float getY() const;
         Vector2 getPosition() const;
-        double getRotation() const;
+        float getRotation() const;
         const Vector2& getScale() const;
-        void setPosition(double x, double y);
+        void setPosition(float x, float y);
         void setPosition(const Vector2& position);
-        void setRotation(double rotation);
-        void setScale(double x, double y);
+        void setRotation(float rotation);
+        void setScale(float x, float y);
         void setScale(const Vector2& scale);
         virtual void render() const = 0;
 
     protected:
         Vector2 position;
-        double rotation = 0.0;
-        Vector2 scale = Vector2(1.0, 1.0);
+        float rotation = 0.0f;
+        Vector2 scale = Vector2(1.0f, 1.0f);
 
     private:
