+++ include/vector.h
@@ -3,9 +3,9 @@
 class Vector2 {
 public:
     Vector2();
-    Vector2(double x, double y);
-    double x = 0;
-    double y = 0;
+    Vector2(float x, float y);
+    float x = 0;
+    float y = 0;
 
 private:
 };
