+++ src/rectangle.cpp
@@ -4,16 +4,16 @@
 
 namespace glvis {
 
-    Rectangle::Rectangle(double width, double height) {
+    Rectangle::Rectangle(float width, float height) {
         this->width = width;
         this->height = height;
 
         const float quadVertices[] = {
             // positions                 // texture Coords
-            (float)(-width / 2.0), (float)( height / 2.0), 0.0f, 0.0f, 1.0f,
-            (float)(-width / 2.0), (float)(-height / 2.0), 0.0f, 0.0f, 0.0f,
-            (float)( width / 2.0), (float)( height / 2.0), 0.0f, 1.0f, 1.0f,
-            (float)( width / 2.0), (float)(-height / 2.0), 0.0f, 1.0f, 0.0f
+            -width / 2.0f,  height / 2.0f, 0.0f, 0.0f, 1.0f,
+            -width / 2.0f, -height / 2.0f, 0.0f, 0.0f, 0.0f,
+             width / 2.0f,  height / 2.0f, 0.0f, 1.0f, 1.0f,
+             width / 2.0f, -height / 2.0f, 0.0f, 1.0f, 0.0f
         };
 
         const unsigned int quadIndices[] = {
@@ -37,11 +37,11 @@ namespace glvis {
         glBindVertexArray(0);
     }
 
-    double Rectangle::getWidth() const {
+    float Rectangle::getWidth() const {
         return width;
     }
 
-    double Rectangle::getHeight() const {
+    float Rectangle::getHeight() const {
         return height;
     }
